refactor: Move shared types into util folder with section visibility type
