+++ packages/ui-app/src/InputExtrinsic/options/method.tsx
@@ -3,13 +3,13 @@
 // of the ISC license. See the LICENSE file for details.
 
 import { Extrinsic$Sections } from '@polkadot/extrinsics/types';
-import { DropdownOptions } from '../types';
+import { DropdownOptions, SectionVisibilityAll } from '../../util/types';
 
 import React from 'react';
 
 import map from '@polkadot/extrinsics';
 
-export default function createOptions (sectionName: Extrinsic$Sections, type: 'private' | 'public'): DropdownOptions {
+export default function createOptions (sectionName: Extrinsic$Sections, type: SectionVisibilityAll): DropdownOptions {
   const section = map[sectionName];
 
   if (!section) {
