+++ packages/ui-app/src/InputExtrinsic/options/section.ts
@@ -3,12 +3,12 @@
 // of the ISC license. See the LICENSE file for details.
 
 import { Extrinsic$Sections } from '@polkadot/extrinsics/types';
-import { DropdownOptions } from '../types';
+import { DropdownOptions, SectionVisibilityAll } from '../../util/types';
 import { shouldDisplaySection } from '../../util/shouldDisplaySection';
 
 import map from '@polkadot/extrinsics';
 
-export default function createOptions (type: 'private' | 'public'): DropdownOptions {
+export default function createOptions (type: SectionVisibilityAll): DropdownOptions {
   return Object
     .keys(map)
     .sort()
