+++ packages/ui-app/src/InputStorage/options/section.ts
@@ -3,7 +3,8 @@
 // of the ISC license. See the LICENSE file for details.
 
 import { Storage$Sections } from '@polkadot/storage/types';
-import { DropdownOptions } from '../../InputExtrinsic/types';
+import { DropdownOptions } from '../../util/types';
+
 import { shouldDisplaySection } from '../../util/shouldDisplaySection';
 
 import map from '@polkadot/storage';
